Sprite drawing nearly done?

+++ BabelEngine4/Assets/Aseprite/AsepriteFrame.cs
@@ -14,6 +14,6 @@ namespace BabelEngine4.Assets.Aseprite
 
         public bool rotated, trimmed;
 
-        public float duration;
+        public int duration;
     }
 }
